Factored out openFile() method from ArbitraryFileWidget to MediaUtils class
